+++ Messages/FIX50/TradeCaptureReportRequest.cs
@@ -1911,36 +1911,36 @@ namespace QuickFix
                 return IsSetField(Tags.InterestAccrualDate);
             }
 
-            public QuickFix.Fields.SecurityStat SecurityStat
+            public QuickFix.Fields.SecurityStatus SecurityStatus
             {
                 get
                 {
-                    QuickFix.Fields.SecurityStat val = new QuickFix.Fields.SecurityStat();
+                    QuickFix.Fields.SecurityStatus val = new QuickFix.Fields.SecurityStatus();
                     GetField(val);
                     return val;
                 }
                 set { SetField(value); }
             }
 
-            public void Set(QuickFix.Fields.SecurityStat val)
+            public void Set(QuickFix.Fields.SecurityStatus val)
             {
-                this.SecurityStat = val;
+                this.SecurityStatus = val;
             }
 
-            public QuickFix.Fields.SecurityStat Get(QuickFix.Fields.SecurityStat val)
+            public QuickFix.Fields.SecurityStatus Get(QuickFix.Fields.SecurityStatus val)
             {
                 GetField(val);
                 return val;
             }
 
-            public bool IsSet(QuickFix.Fields.SecurityStat val)
+            public bool IsSet(QuickFix.Fields.SecurityStatus val)
             {
-                return IsSetSecurityStat();
+                return IsSetSecurityStatus();
             }
 
-            public bool IsSetSecurityStat()
+            public bool IsSetSecurityStatus()
             {
-                return IsSetField(Tags.SecurityStat);
+                return IsSetField(Tags.SecurityStatus);
             }
 
             public QuickFix.Fields.SettleOnOpenFlag SettleOnOpenFlag